+++ ui/components/app/confirm/info/row/index.ts
@@ -3,6 +3,6 @@ export * from './date';
 export * from './divider';
 export * from './row';
 export * from './text';
-export * from './text-token';
+export * from './text-token-units';
 export * from './url';
 export * from './value-double';
